FLU changes for 64-bit compile errors

+++ src/flu/Flu_File_Chooser.cpp
@@ -2773,7 +2773,7 @@ int Flu_File_Chooser :: popupContextMenu( Entry *entry )
   const Fl_Menu_Item *selection = entryPopup.popup();
   if( selection )
     {
-      int handler = (int)selection->user_data();
+      long handler = (long)selection->user_data();
       switch( handler )
 	{
 	case ACTION_NEW_FOLDER:
